+++ packages/x-components/src/x-modules/facets-next/entities/__tests__/hierarchical-filter-entity.spec.ts
@@ -1,36 +1,14 @@
 import { Filter } from '@empathyco/x-types-next';
-import { createLocalVue } from '@vue/test-utils';
-import Vuex, { Store } from 'vuex';
+import { Store } from 'vuex';
 import { createNextHierarchicalFacetStub } from '../../../../__stubs__/facets-stubs.factory';
 import { createNextHierarchicalFilter } from '../../../../__stubs__/filters-stubs.factory';
-import { resetStoreXModuleState } from '../../../../__tests__/utils';
 import { RootXStoreState } from '../../../../store/store.types';
-import { arrayToObject } from '../../../../utils/array';
-import { facetsNextXStoreModule } from '../../store/module';
 import { HierarchicalFilterEntity } from '../hierarchical-filter.entity';
+import { prepareFacetsStoreWithFilters } from './utils';
 
 describe('testing HierarchicalFilterEntity', () => {
-  function prepareFacetsStore(filters: Filter[]): Store<RootXStoreState> {
-    const vue = createLocalVue();
-    vue.use(Vuex);
-    const store = new Store({
-      modules: {
-        x: {
-          modules: {
-            facetsNext: { ...facetsNextXStoreModule, namespaced: true }
-          },
-          namespaced: true
-        }
-      }
-    });
-    resetStoreXModuleState(store, 'facetsNext', facetsNextXStoreModule.state(), {
-      filters: arrayToObject(filters, 'id')
-    });
-    return store;
-  }
-
   it('allows selecting a filter that is not in the store', () => {
-    const store = prepareFacetsStore([]);
+    const store = prepareFacetsStoreWithFilters([]);
     const categoryWomen = createNextHierarchicalFilter('category', 'women', false);
     const categoryWomenEntity = new HierarchicalFilterEntity(store, categoryWomen);
 
@@ -39,7 +17,7 @@ describe('testing HierarchicalFilterEntity', () => {
   });
 
   it('allows deselecting a filter that is not in the store', () => {
-    const store = prepareFacetsStore([]);
+    const store = prepareFacetsStoreWithFilters([]);
     const categoryWomen = createNextHierarchicalFilter('category', 'women', true);
     const categoryWomenEntity = new HierarchicalFilterEntity(store, categoryWomen);
 
@@ -49,11 +27,11 @@ describe('testing HierarchicalFilterEntity', () => {
 
   it('allows selecting and deselecting a filter', () => {
     /*
-     Men
-      Shirts
-        Long Sleeve
-      Jeans
      Women
+     Men
+       Jeans
+       Shirts
+         Long Sleeve
      */
     const categoryFacet = createNextHierarchicalFacetStub('category', createChild => [
       ...createChild('men', false, createChild => [
@@ -62,7 +40,7 @@ describe('testing HierarchicalFilterEntity', () => {
       ]),
       ...createChild('women', false)
     ]);
-    const store = prepareFacetsStore(categoryFacet.filters);
+    const store = prepareFacetsStoreWithFilters(categoryFacet.filters);
     const [categoryMen, categoryShirts, categoryLongSleeve, categoryJeans, categoryWomen] =
       categoryFacet.filters;
     const categoryLongSleeveEntity = new HierarchicalFilterEntity(store, categoryLongSleeve);
